# copied-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ZenChattyServer.Net",
    "version": "1.0"
  },
  "paths": {
...
    "/api/social/get-user-info/{targetSeekBy}": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetSeekBy",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/private-chat": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrivateChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrivateChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrivateChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/is-disabled/{targetUserId}": {
      "get": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/block/{targetUserId}": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "block",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/unblock-and-add/{targetUserId}": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/add-friend/{targetUserGuid}": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetUserGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/group/create": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/group/leave-from/{groupId}": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/group/get-announcements/{groupId}": {
      "get": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/social/is-blocked/{targetUserId}": {
      "get": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/social/contact/get-unread/{contactId}": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unreadCount",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/social/privacy/update": {
      "patch": {
        "tags": [
          "UserSocial"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacySettings"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacySettings"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PrivacySettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/social/add-friend-request": {
      "post": {
        "tags": [
          "UserSocial"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "viaGroupId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BasicResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ChatResponse": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "nullable": true
          },
          "chatType": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "unreadCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "isPinned": {
            "type": "boolean"
          },
          "lastMessagePreview": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateGroupChatRequest": {
        "required": [
          "groupName"
        ],
        "type": "object",
        "properties": {
          "groupName": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "allowMemberInvite": {
            "type": "boolean"
          },
          "requireApproval": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CreatePrivateChatRequest": {
        "required": [
          "receiverId"
        ],
        "type": "object",
        "properties": {
          "receiverId": {
            "minLength": 1,
            "type": "string"
          },
          "isInformal": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EGender": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "isCanceled": {
            "type": "boolean"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "info": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/EMessageType"
          },
          "sentTimestamp": {
            "type": "integer",
            "format": "int64"
          },
          "serverCaughtTimestamp": {
            "type": "integer",
            "format": "int64"
          },
          "isMentioningAll": {
            "type": "boolean"
          },
          "mentionedUserGuids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isAnnouncement": {
            "type": "boolean"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "sender": {
            "$ref": "#/components/schemas/User"
          },
          "viaGroupChatId": {
            "type": "string",
            "nullable": true
          },
          "ofChatId": {
            "type": "string",
            "nullable": true
          },
          "ofChat": {
            "$ref": "#/components/schemas/Chat"
          }
        },
        "additionalProperties": false
      },
      "PrivacySettings": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "isDiscoverableViaSearch": {
            "type": "boolean"
          },
          "isInvitableToGroup": {
            "type": "boolean"
          },
          "isAddableFromGroup": {
            "type": "boolean"
          },
          "isNewChatKeepSilent": {
            "type": "boolean"
          },
          "allowViewInfoFromGroupChat": {
            "type": "boolean"
          },
          "contactVisibility": {
            "$ref": "#/components/schemas/EPrivacyVisibilityRange"
          },
          "bioVisibility": {
            "$ref": "#/components/schemas/EPrivacyVisibilityRange"
          },
          "genderVisibility": {
            "$ref": "#/components/schemas/EPrivacyVisibilityRange"
          },
          "birthdayVisibility": {
            "$ref": "#/components/schemas/EPrivacyVisibilityRange"
          }
        },
        "additionalProperties": false
      },
...
      "UserInfoResponse": {
        "type": "object",
        "properties": {
          "localId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarFileLocator": {
            "type": "string",
            "nullable": true
          },
          "backgroundFileLocator": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "customId": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EUserStatus"
          },
          "gender": {
            "$ref": "#/components/schemas/EGender"
          },
          "birth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}